Extract customer detail cards into their own components
The customer detail route had grown to 1500 lines holding nine cards and
their dialogs. Each card now lives in its own file alongside the other
resource components, leaving the route with just the page shell.

No behaviour change.
